rewrote backend with extensible endpoints
still using hard-coded playlist as a test case, because I can't get apple music api yet.

Now have a /api/commonplaylistobject endpoint so that we can create and store playlists from any origin or api we may want to support.

And a separate, stateless endpoint /api/spotifyplaylist to add a commonplaylistobject to spotify library. We may also implement endpoints like /api/ytmusicplaylist with the same functionality, subbing in a different API.

Much neater.

User may now log in with spotify.

Access keys are generated on an as-needed basis to reduce api calls.